backend/requirements.txt

flask
flask-cors
requests
google-genai
supabase
pytz
google-cloud-aiplatform
